requirements.txt
numpy==1.26.4
ogb==1.3.6
pydantic==2.7.1
PyYAML==6.0.1
scipy==1.13.0
tensorboardX==2.6.2.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1+cu118
--find-links https://data.pyg.org/whl/torch-2.2.1+cu118.html
torch_sparse==0.6.18+pt22cu118
--find-links https://data.pyg.org/whl/torch-2.2.1+cu118.html
torch_scatter==2.1.2+pt22cu118
torch_geometric==2.5.3
torchtyping==0.1.4
tqdm==4.66.2
typeguard==4.2.1